### 2021-03-23 #3130 @Kevin-Mattheus-Moerman
@felipeZ thank you for this submission. I have just flagged this work for additional scope review by our editorial board. This is typical for works presenting with <1000 lines of code. We will review if this work conforms to our  ["substantial scholarly effort" requirements](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). The scope review should take about a week.
