+++ README.rst
@@ -25,7 +25,7 @@ See `documentation <https://ceiba.readthedocs.io/en/latest/>`_ and `blog post <h
 
 Installation
 ************
-The [provisioning folder](https://github.com/nlesc-nano/ceiba/tree/main/provisioning) contains the instructions
+The `provisioning folder <https://github.com/nlesc-nano/ceiba/tree/main/provisioning>`_ contains the instructions
 to deploy the web service using docker. Alternatively, you can deploy the web service locally using
 the following instructions:
 
