### 2021-04-02 #3130 @kthyng
@felipeZ I think the paper is not generating because of the tab highlighted in red here: https://github.com/nlesc-nano/ceiba/blob/main/paper/paper.md. PLease fix that.

The editorial board has decided this is in scope, when reviewed together with the CLI package.
