### 2021-04-02 #3130 @majensen
Looks like an interesting approach, happy to handle this
