+++ provisioning/README.md
@@ -3,7 +3,7 @@ This recipe installs and configures the **ceiba** app using [ansible](https://ww
 
 ## Step to install the app
 1. Install [Ansible](https://docs.ansible.com/ansible/latest/installation_guide/intro_installation.html) in your computer.
-2. Clone [this repo](https://github.com/NLESC-JCER/gitlab_runner).
+2. Clone [this repo](https://github.com/nlesc-nano/ceiba)
 3. Edit the [inventory](https://docs.ansible.com/ansible/latest/user_guide/intro_inventory.html) file with the address of the server(s) where you want to install the runner.
 4. Edit the [playbook](https://docs.ansible.com/ansible/latest/user_guide/playbooks.html) file with the `remote_user` name for the hosts.
 5. Make sure that you can ssh to your server(s).
