### 2021-04-06 #3130 @majensen
Hi @vsoch -- would you have time and inclination to review [this work](#3130) for [JOSS](https://joss.theoj.org)? Your expertise and experience would be very much appreciated!
Thanks for considering --@majensen
