### 2021-04-20 #3197 @majensen
Hi @felipeZ - whedon has a suggestion for a DOI for the SLURM reference (see https://github.com/openjournals/joss-reviews/issues/3197#issuecomment-823572123) - could you add to your bib when you have an opportunity?
