### 2021-04-21 8442ff6
add doi to slurm reference
