+++ paper/paper.bib
@@ -91,7 +91,8 @@ articleno = {2}
     booktitle = {In Lecture Notes in Computer Science: Proceedings of Job Scheduling Strategies for Parallel Processing (JSSPP) 2003},
     year = {2002},
     pages = {44--60},
-    publisher = {Springer-Verlag}
+    publisher = {Springer-Verlag},
+    doi = {10.1007/10968987_3}
 }
 
 @misc{Proteins_benchmark,
