### 2021-05-24 nlesc-nano/ceiba#30 @junjun-zhang
Missing documentation on how to add authenticated users
When running the `ceibacli login` command using the GitHub PAT generated from my account, I got an error message: `User `junjun-zhang` doesn't have permissions to access the service`. Checked the ceiba code and noticed that there is a Mongo DB collection `authenticated_users`, however, there seems to be no documentation on how to add users to the collection.
