### 2021-05-24 #3197 @junjun-zhang
Have noticed some differences between the instructions described in the **Examples** section of the paper and that of the Ceiba documentation website.

I would suggest updating the documentation site to the latest and greatest since there will be where Ceiba users are expected to use. Also, checked a few other papers published on JOSS, they don’t typically include detailed step-by-step instructions. It's probably better to leave those in the software tool’s documentation and the paper only refers to them as needed.

Specific findings so far:
1. I was able to get the server up and running using `./start_app.sh` command, maybe there should be a script to stop the server, eg, `stop_app.sh`
2. Problem with `ceiba-cli` installation, reported [here](https://github.com/nlesc-nano/ceiba-cli/issues/30). To continue my testing, I run the ceiba-cli tool from the source code folder by `python -m ceibacli.cli`
3. Problem with `login` as reported [here](https://github.com/nlesc-nano/ceiba/issues/30)

Will continue my evaluation when the third issue is resolved.
