### 2021-05-25 9ae4f06
fix #30 by adding the docs to add users to the web service
