### 2021-05-27 nlesc-nano/ceiba#30 @junjun-zhang
Thanks @felipeZ for the quick response, however, the problem persists even I edited the `users.txt` file.

Had a look at the docker compose file, based on this line:
`entrypoint: ceiba -m 172.17.0.2 -u admin -p ${MONGO_PASSWORD} -f users.txt`

The `users.txt` file has to be within the docker contain, not the one from the host. Basically it would require the docker image to be rebuilt in order to include the updated `users.txt` included in the container. Alternatively, a volume can be mount that include this file.
