### 2021-05-27 #3197 @junjun-zhang
@felipeZ would be nice to have this addressed: https://github.com/nlesc-nano/ceiba/issues/16. Ideally users can be added / removed via API endpoints
