### 2021-05-31 #3197 @junjun-zhang
@felipeZ any suggestions on how I may be able to continue testing ceiba? Maybe before https://github.com/nlesc-nano/ceiba/issues/16 is addressed, you could rebuild the docker image to add my username `junjun-zhang` (also add @dimitridarras ?) to the users.txt file? Not ideal, but it would allow us to continue the review process.
