### 2021-05-31 ace2805
add Junjun and Dimitri to the available users to the docker container
