### 2021-06-21 #3197 @junjun-zhang
I was able to set up my own Ceiba server after locally building a new Docker image and tagging it with `ghcr.io/nlesc-nano/ceiba:prototype`. The docker image pulled from GHCR does not work. With that I continued the testing with different use cases. The overall experience working with Ceiba and Ceiba-CLI is good, in general there is no problem following the documentation. Below are findings and suggestions:

**The paper:**
- Suggest to split the current “Statement of need” section into “Statement of need” (only include the first paragraph) and “Implementation” sections.
- Remove the “Example” section from the paper, instead, incorporate the content of “Example” to relevant sections of the Ceiba and Ceiba-CLI documentation sites.
- Adjustment to Figure 1: the diagram is a bit misleading, Ceiba-CLI client does not actually interact with database directly, database updates initiated from the client are always done via Ceiba Server. The diagram should reflect this, maybe move the Database box behind the Server. The current blue box at the center does not represent any abstraction or concept, and should therefore be removed. Also, the box with text `Web Service Functional Requirements` seems not quite relevant and does not contribute any value.

**Ceiba Server:**
- Instead of hardcoding user accounts, it is strongly suggested to complete the proper user account management feature (https://github.com/nlesc-nano/ceiba/issues/16). This would greatly improve usability and allow more users to adopt Ceiba. Service should include add / remove (maybe disable instead of remove?) normal users and admin users. Ceiba-CLI may accordingly add user account management capability as well.
- Is there a way to prevent duplicate jobs from being added?
- It seems there is no query function for job status
  - Can admins see all jobs in different statuses: running, reserved, completed etc?
  - Can an individual user see the status of her own job?
  - It would be helpful if the admin is able to see which jobs are reserved by which users, this way, the admin would be able to get in touch with the user for any possible followup inquiries.

**Ceiba-CLI Client:**
- User account management as mentioned in the Server section.
- Suggest to display acknowledgement upon user’s successful login.
- Would be nice to have a logout command.
- Currently, there is no direct support from Ceiba / Ceiba-CLI to monitor on-going jobs. This can present challenges when there are many running or pending jobs. When large number of jobs failed, how to restart them again? Maybe the author can suggest some good practices.
- When I followed the step to run the `compute` command, I got `/bin/sh: computepi.py: command not found` error. It seems like the value of `command` in compute_input.yml needs to include a full path.
- Related to the `compute` command, currently Ceiba does not provide any way to manage the code of the actual compute, be it a script or executable binary, or anything else. Is it possible to register those as well at the Ceiba Server, similar to registering job parameters? For example, the `computepi.py` script, can a git url pointing to that script be added to the ‘collection’? This way, when Ceiba-CLI fetches job parameters, it can also fetch the script. For complex application, there may be some way to instruction how it can be installed and run.
- Would be nice to implement auto reporting. It could be as an additional command after the original compute completed successfully.
- For reporting large data objects, the documentation suggested using SWIFT object store from OpenStack, but also warned it costs money. Would minio (https://github.com/minio/minio) be a suitable solution, assuming users are able to set up and maintain minio?
