### 2021-12-27 #3197 @dimitridarras
Apologies for the late response, I need to bow out from the review as well, but thought I'd add some comments from my initial build  almost a year ago that may or may not be relevant.   Please take with a huge grain of salt as this was written a while back while flying blind!   I do not have a scientific background  - IT and web.

Installation comments:

Be explicit in Installation instructions - i.e. these are the full commands for mac

nano  ~/.bash_profile
#vi  ~/.bash_profileœ
export MONGO_PASSWORD='123' source ~/.bash_profile echo $MONGO_PASSWORD

On two different macs, received the same shell/python-related error that was corrected only by identifying a lurking environment variable:

TempLock:provisioning focus$ ./start_app.sh ERROR: .IsADirectoryError:
[Errno 21] Is a directory: './' ERROR: .IsADirectoryError: [Errno 21] Is
a directory: './'

Unless I'm missing something, your shell script docker-compose.yml reference is not discussed in the instructions.   To serve the user better and have a turnkey developer experience and strive for clarity, please sub "docker-compose.yml" for {DOCKER_COMPOSE_FILE} or define the env var.

8	docker-compose -f "${DOCKER_COMPOSE_FILE}" pull 9	docker-compose
-f "${DOCKER_COMPOSE_FILE}" up -d

Accessing Docker instance instructions could be helpful - ssh credentials, etc.  I'm using Docker Desktop and the mac terminal to access them.  I can _see_ them no problem but I'm just wondering in a true distributed workflow how someone in say the Netherlands and someone in South Africa would interface with the system?   I'm guessing hosted cloud instances and keys/SSH?  These details may seem obvious but addressing them would allow the paper to shine in scenarios where end users have gaps in knowledge or limited infrastructure experience.

Blog and Paper Comments:

Blog post presents more accessibly than the paper although both are missing content and need to be fleshed-out out further.   The paper hints at large-scale enterprise computations but doesn't actually go there.

I like the reference to the following at home project (https://foldingathome.org/) - why is it in the blog post only?

Why is the guide "by no means a complete guide to writing web applications?" as mentioned in the blog post?  I think more specificity is needed here - i.e. the stack and libraries and LoE (Level of Effort) to customize the framework you've put in place.   Is the paper presenting a methodology, a stack, a software framework, or all?  How do you provision for updates and version control for the underlying software stack, i.e. Mongo compatibility in the future with Python's GraphQL software?

System interaction:

I'm not getting the positional arguments seen in the blog post with:

	ceiba --help

I have the system up and working and can check the baseline functionality requirement, however I really don't know what to do next.

Overall comments and requested additional information:

Documentation is scant and needs to be fleshed out more.  I can go into details but feel like it needs more substance and direction even if "obvious".

**This collection/stack needs an illustrative example for the reviewer to actually work through it.  I thought I saw it somewhere but then was not able to locate it.  Even if it's a series
of steps or a tutorial, having concrete instructions will illustrate the process better than having me fill in the blanks.   Better yet, consider multiple scenarios such as:

A) Your personal data management, pending permissions and visibility
B) The Monte Carlo example - high and low computational needs (scenario A - localhost with Docker specs maxed out, scenario B, cloud cluster)
C) Domain specific illustrations from say fluid dynamics and life sciences to get a sense of how the service handles differing schemata
D) Dummy data

In on of my employer's projects, we use illustrative datasets to rinse through the system and evidence functionality and display.  It's not authoritative but it gives a flavor of how the database, queries, and views work.  Setting up the container with an example database to my mind solves this issue of "how does it work?".   If at all possible, try to avoid having me load the database itself, but in a pinch that's better than nothing.

The Pi/Monte Carlo example seemed promising - why is that json file not in the repo?  How do I run it? Furthermore, there's an indication that the system will allow for computationally expensive transactions, however, there are no benchmarks or hardware/virtualized hardware specifications to back this up in real life.   For instance:  I'm running a simulation using
python and a relational database on AWS and it cost me $400 versus significantly less with Ceiba...

I like the concluding comments for the system administrator Adam who is setting up the cloud instance and will do it myself as it is a requirement for acceptance, to my mind since a local installation is not indicative of your true end-user scenario.

It's worth noting for the novice user/administrator that you are not providing a UI and all project work will be CLI-based.   Do you have a recommendation for the project administrator for implementing a UI or is the assumption that end users have expertise with the cli and Docker?
