### 2022-02-03 c89a96f
Update CITATION.cff cffversion to 1.2.0
