### 2022-02-03 916992d
Fix license in CITATION.cff
