+++ CITATION.cff
@@ -8,10 +8,10 @@ authors:
     given-names: Felipe
     orcid: "https://orcid.org/0000-0001-8286-677X"
 date-released: 2020-10-03
-keywords: 
+keywords:
   - "quantum-chemistry"
   - "materials-science"
 version: "0.1.0"
 repository-code: "https://github.com/nlesc-nano/ceiba"
 message: "If you use this software, please cite it using these metadata."
-license: "Apache Software License 2.0"
+license: Apache-2.0
