### 2022-06-28 #3197 @majensen
@openjournals/joss-eics, the author and I agree that a withdrawal of this paper should proceed. Can you assist? Thanks.
